requirements.txt
transformers>=4.41.0
accelerate>=0.31.0
datasets>=2.20.0
sentencepiece>=0.2.0
sacrebleu>=2.4.2
textstat>=0.7.4
peft>=0.12.0
bitsandbytes>=0.43.1; platform_system != "Windows"
mlflow>=2.14.3
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
jinja2>=3.1.4
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
pyyaml>=6.0.1
